Ensure reconstruction initiation uses new pipeline prep
