+++ ServiceLayer/BlockFemReconstructionService.cs
@@ -34,10 +34,6 @@ namespace ServiceLayer
         private bool _initialized;
         // Monotonically increasing index of frames emitted so far (across cycles)
         private int _frameIndex;
-        // Iteration counter used to throttle expensive error/metric calculations
-        private int _iterationCount;
-        // Buffer to accumulate frames of the current drive-pattern cycle; flushed into a result when the cycle completes
-        private readonly List<ReconstructionFrame> _cycleFrames = new();
 
         /// <inheritdoc />
         public event EventHandler<ReconstructionResult>? ReconstructionUpdated;
@@ -104,9 +100,7 @@ namespace ServiceLayer
             // 9) Reset UI-facing collections and internal buffers for a fresh session.
             Workspace.SetReconstructionFrames(new List<ReconstructionFrame>());
             Workspace.SetReconstructionResults(new List<ReconstructionResult>());
-            _cycleFrames.Clear();
             _frameIndex = 0;
-            _iterationCount = 0;
             _initialized = true;
         }
 
@@ -116,17 +110,7 @@ namespace ServiceLayer
                                                                            double excitationAmplitude)
             => Task.Run(() =>
             {
-                // Execute exactly one complete drive-pattern cycle of reconstruction.
-                // Each call to ExecuteReconstructionStep processes one measurement frame and may emit a result
-                // only when the cycle completes. We return the last non-null result from the loop.
-                ReconstructionResult? result = null;
-                int cycleLength = Math.Max(1, _measurementService.FramesPerCycle);
-                for (int i = 0; i < cycleLength; i++)
-                {
-                    result = ExecuteReconstructionStep(stepSize, regularizationWeight, excitationAmplitude);
-                }
-
-                return result;
+                return ExecuteReconstructionStep(stepSize, regularizationWeight, excitationAmplitude);
             });
 
         /// <inheritdoc />
@@ -153,57 +137,16 @@ namespace ServiceLayer
                 //    and stores the drive-pattern step index for downstream boundary condition reconstruction.
                 var measurement = PrepareMeasurement(excitationAmplitude);
 
-                // 2) Execute the reconstruction step in the persistence layer. This returns one
-                //    ReconstructionFrame per provided measurement frame (usually 1 here).
-                var frames = _persistence.Step(measurement, _frameIndex);
+                var result = _persistence.RunCycle(measurement, stepSize, regularizationWeight);
 
-                // 3) Surface each frame to the workspace and UI, and buffer them for the current cycle.
-                foreach (var frame in frames)
+                foreach (var frame in result.Frames)
                 {
-                    _cycleFrames.Add(frame);
                     PublishFrame(frame);
                 }
 
-                // 4) Advance the global frame counter by the number of frames produced.
-                _frameIndex += frames.Count;
-
-                // 5) If the current cycle is not complete yet, we're done for this step (no aggregated result yet).
-                //if (_frameIndex % Math.Max(1, _measurementService.FramesPerCycle) != 0)
-                //    return null;
-
-                // 6) On cycle completion, assemble and apply a gradient-based conductivity update and publish a result.
-                var mesh = _runtimeContext.RuntimeMesh
-                          ?? throw new InvalidOperationException("Runtime mesh missing from reconstruction context.");
-
-                // Snapshot the previous estimate to include it in the result payload.
-                var previous = mesh.GetConductivityDistribution();
-
-                // Apply optimizer(s) with regularization to get the updated conductivity.
-                var updated = ApplyGradientUpdate(_cycleFrames, previous, stepSize, regularizationWeight);
-                if (updated == null)
-                    return null;
-
-                // Build a result that captures the discretization and the cycle's frames for UI/export.
-                var result = new ReconstructionResult(mesh.GetDiscretization(),
-                                                      _runtimeContext.OriginalDistribution,
-                                                      previous,
-                                                      updated,
-                                                      new List<ReconstructionFrame>(_cycleFrames));
-
-                // Persist for UI consumption and clear the buffer for the next cycle.
-                Workspace.AddReconstructionResultToWorkspace(result);
-                _cycleFrames.Clear();
-
-                // Only surface expensive error/metric calculations every 10th iteration
-                // to match the behaviour of the classic reconstruction pipeline.
-                _iterationCount++;
-                if (_iterationCount % 10 == 0)
-                {
-                    ReconstructionUpdated?.Invoke(this, result);
-                    return result;
-                }
-
-                return null;
+                _frameIndex += result.Frames.Count;
+                ReconstructionUpdated?.Invoke(this, result);
+                return result;
             }
             catch (Exception ex)
             {
@@ -270,181 +213,6 @@ namespace ServiceLayer
             };
 
             return measurement;
-        }
-
-        // Applies an optimization update over the accumulated frames of the current cycle.
-        // 1) Aggregate optimizer gradients and regularizations across frames.
-        // 2) Apply configured numeric optimizer(s) to produce an updated conductivity.
-        // 3) Clip the conductivity to physically plausible bounds.
-        // 4) Inform persistence and workspace of the new baseline for the next cycle.
-        private ConductivityDistribution? ApplyGradientUpdate(IReadOnlyList<ReconstructionFrame> frames,
-                                                              ConductivityDistribution currentSigma,
-                                                              double stepSize,
-                                                              double regularizationWeight)
-        {
-            if (_runtimeContext == null)
-                return null;
-            if (frames.Count == 0)
-                return null;
-
-            // Combine per-frame optimizer gradients and regularizations into a single averaged gradient per optimizer id.
-            var optimizerGradients = AggregateOptimizerGradients(frames, regularizationWeight);
-
-            // Run one optimization step using the configured numeric optimizer(s).
-            var updated = ApplyOptimizers(currentSigma, optimizerGradients, stepSize);
-
-            // Keep the estimate in reasonable bounds (implementation-specific policy in the clipper).
-            updated = ConductivityClipper.Clip(updated);
-
-            // Synchronize the internal state in persistence and workspace so that subsequent
-            // steps compute regularization/gradients with respect to the latest estimate.
-            _persistence.UpdateCurrentDistribution(updated);
-            Workspace.SetInitialConductivityDistribution(updated);
-            return updated;
-        }
-
-        // Aggregates optimizer gradients and regularization contributions across the frames of one cycle.
-        // The result is a per-optimizer-id ConductivityDistribution representing
-        //    (Sum(?J/??) - ? * Sum(?R/??)) / N_frames
-        // where ? is the provided regularizationWeight.
-        private static Dictionary<string, ConductivityDistribution> AggregateOptimizerGradients(IReadOnlyList<ReconstructionFrame> frames,
-                                                                                               double regularizationWeight)
-        {
-            // Temporary accumulation maps keyed by optimizer id; inner map is elementId -> value.
-            var gradientAccum = new Dictionary<string, Dictionary<int, double>>();
-            var regAccum = new Dictionary<string, Dictionary<int, double>>();
-
-            foreach (var frame in frames)
-            {
-                // Sum pure data-misfit gradients per optimizer id
-                foreach (var gradientEntry in frame.OptimizerGradients)
-                {
-                    if (!gradientAccum.TryGetValue(gradientEntry.Key, out var dict))
-                    {
-                        dict = new Dictionary<int, double>();
-                        gradientAccum[gradientEntry.Key] = dict;
-                    }
-
-                    foreach (var kvp in gradientEntry.Value.Conductivities)
-                    {
-                        dict[kvp.Key] = dict.TryGetValue(kvp.Key, out var existing)
-                            ? existing + kvp.Value
-                            : kvp.Value;
-                    }
-                }
-
-                // Sum regularization gradients per optimizer id
-                foreach (var regEntry in frame.OptimizerRegularizations)
-                {
-                    if (!regAccum.TryGetValue(regEntry.Key, out var dict))
-                    {
-                        dict = new Dictionary<int, double>();
-                        regAccum[regEntry.Key] = dict;
-                    }
-
-                    foreach (var kvp in regEntry.Value.Conductivities)
-                    {
-                        dict[kvp.Key] = dict.TryGetValue(kvp.Key, out var existing)
-                            ? existing + kvp.Value
-                            : kvp.Value;
-                    }
-                }
-            }
-
-            // Average over the number of frames to get a stable per-cycle gradient.
-            int frameCount = Math.Max(1, frames.Count);
-            var result = new Dictionary<string, ConductivityDistribution>();
-
-            // Merge keys from both accumulators since an optimizer id may appear in only one.
-            foreach (var optimizerId in gradientAccum.Keys.Union(regAccum.Keys))
-            {
-                var gradDict = gradientAccum.TryGetValue(optimizerId, out var g) ? g : new Dictionary<int, double>();
-                var regDict = regAccum.TryGetValue(optimizerId, out var r) ? r : new Dictionary<int, double>();
-
-                var combined = new Dictionary<int, double>();
-                foreach (var kvp in gradDict)
-                {
-                    double reg = regDict.TryGetValue(kvp.Key, out var regVal) ? regVal : 0.0;
-                    // Combine misfit gradient and regularization with ? and average across frames.
-                    combined[kvp.Key] = (kvp.Value - regularizationWeight * reg) / frameCount;
-                }
-
-                // Include entries that exist only in the regularization map.
-                foreach (var kvp in regDict)
-                {
-                    if (combined.ContainsKey(kvp.Key))
-                        continue;
-                    combined[kvp.Key] = (-regularizationWeight * kvp.Value) / frameCount;
-                }
-
-                result[optimizerId] = new ConductivityDistribution(combined);
-            }
-
-            return result;
-        }
-
-        // Applies one optimization step with the configured numeric optimizer(s).
-        // - If only one optimizer is present, apply it directly to currentSigma using its gradient.
-        // - If multiple optimizers exist, compute each candidate update and return the weighted average by configured weights.
-        private ConductivityDistribution ApplyOptimizers(ConductivityDistribution currentSigma,
-                                                         IReadOnlyDictionary<string, ConductivityDistribution> optimizerGradients,
-                                                         double stepSize)
-        {
-            // If no optimizers are configured, keep the current estimate unchanged.
-            if (_runtimeContext == null || _runtimeContext.NumericOptimizers == null || _runtimeContext.NumericOptimizers.Count == 0)
-                return currentSigma;
-
-            // Fast-path: single optimizer uses the only available gradient (or an empty one if missing).
-            if (_runtimeContext.NumericOptimizers.Count == 1)
-            {
-                var optimizer = _runtimeContext.NumericOptimizers[0];
-                var gradient = optimizerGradients.Values.FirstOrDefault() ?? new ConductivityDistribution(new Dictionary<int, double>());
-                var candidate = optimizer.numericOptimizer.OptimizationStep(currentSigma, gradient, stepSize);
-                return MergeWithBaseline(currentSigma, candidate);
-            }
-
-            // Multiple optimizers case: compute each candidate update and form a weighted average by connection weight.
-            var weightedSum = new Dictionary<int, double>();
-            double totalWeight = 0.0;
-
-            foreach (var (id, weight, optimizer) in _runtimeContext.NumericOptimizers)
-            {
-                var gradient = optimizerGradients.TryGetValue(id, out var specific)
-                    ? specific
-                    : optimizerGradients.Values.FirstOrDefault() ?? new ConductivityDistribution(new Dictionary<int, double>());
-
-                var candidate = optimizer.OptimizationStep(currentSigma, gradient, stepSize);
-                foreach (var kvp in candidate.Conductivities)
-                {
-                    weightedSum[kvp.Key] = weightedSum.TryGetValue(kvp.Key, out var existing)
-                        ? existing + weight * kvp.Value
-                        : weight * kvp.Value;
-                }
-
-                totalWeight += weight;
-            }
-
-            // Guard: if all weights are zero, fall back to identity update.
-            if (totalWeight <= double.Epsilon)
-                return currentSigma;
-
-            // Normalize by the total weight to get the final combined candidate.
-            var combined = weightedSum.ToDictionary(kvp => kvp.Key, kvp => kvp.Value / totalWeight);
-            var combinedDistribution = new ConductivityDistribution(combined);
-            return MergeWithBaseline(currentSigma, combinedDistribution);
-        }
-
-        // Ensures that an updated conductivity distribution preserves any elements that were not explicitly
-        // produced by an optimizer (e.g., sparse updates). Missing keys would otherwise be interpreted as zero
-        // conductivity, causing the UI canvas to diverge from the solver state.
-        private static ConductivityDistribution MergeWithBaseline(ConductivityDistribution baseline,
-                                                                  ConductivityDistribution updated)
-        {
-            var merged = new Dictionary<int, double>(baseline.Conductivities);
-            foreach (var kvp in updated.Conductivities)
-                merged[kvp.Key] = kvp.Value;
-
-            return new ConductivityDistribution(merged);
         }
     }
 }
